requirements.txt
cachelib==0.1.1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
ecdsa==0.14.1
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-Session==0.3.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
python-jose==3.2.0
requests==2.25.0
rsa==4.6
six==1.15.0
urllib3==1.26.2
Werkzeug==1.0.1
